svg map beginnings
