+++ app/assets/stylesheets/general.scss
@@ -1,0 +1,3 @@
+.body {
+  background-color: #F6F6F6;
+}
