+++ app/assets/stylesheets/map.scss
@@ -1,0 +1,4 @@
+.map {
+  width: 80%;
+  margin: 0 auto;
+}
