+++ app/javascript/react/components/StateTile.js
@@ -1,17 +1,17 @@
-import React from "react"
-import { Link } from "react-router-dom"
-
-import StateShowContainer from "./StateShowContainer"
-
-const StateTile = (props) => {
-
-  return (
-    <div className="row">
-      <Link to={`/states/${props.id}`}>
-        <p className="state-name">{props.name}</p>
-      </Link>
-    </div>
-  )
-}
-
-export default StateTile;
+// import React from "react"
+// import { Link } from "react-router-dom"
+//
+// import StateShowContainer from "./StateShowContainer"
+//
+// const StateTile = (props) => {
+//
+//   return (
+//     <div className="row">
+//       <Link to={`/states/${props.id}`}>
+//         <p className="state-name">{props.name}</p>
+//       </Link>
+//     </div>
+//   )
+// }
+//
+// export default StateTile;
