+++ app/views/layouts/application.html.erb
@@ -7,7 +7,7 @@
   <link rel="stylesheet" href="//cdn.jsdelivr.net/npm/semantic-ui@2.4.2/dist/semantic.min.css" />
   <%= csrf_meta_tags %>
 </head>
-<body>
+<body class="body">
 
   <div class="ui menu">
     <h1><a href="/">Home Matters</a></h1>
